Split up existing examples into sub-directories
