+++ example/builder/src/private_component.dart
