+++ example/builder/src/private_factory_public_component.dart
