+++ example/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta http-equiv="X-UA-Compatible" content="IE=edge, chrome=1">
     <meta charset="utf-8">
-    <title>over_react builder testing</title>
+    <title>OverReact Examples</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="apple-mobile-web-app-capable" content="yes">
     <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent">
@@ -15,11 +15,10 @@
 
 </head>
 <body>
-    <div id="content"></div>
-
-    <script src="packages/react/react.js"></script>
-    <script src="packages/react/react_dom.js"></script>
-
-    <script type="application/javascript" defer src="main.dart.js"></script>
+  <h1 class="display-4" style="margin: 2rem 0;">OverReact Examples</h1>
+  <ul>
+    <h3><li><a href="context/">Context Example</a></li></h3>
+    <h3><li><a href="builder/">Builder Testing <em>(for internal use only)</em></a></li></h3>
+  </ul>
 </body>
 </html>
